fix: Make the reclaim-space mode logic more robust
Improve the reclaim-space mode logic and make it more robust:

The previous logic was checking if the closed window was "leftmost"
(meaning it was the first in the viewport) and nudge the focus left /
right if it wasn't. This is fragile and slightly too simplistic because:

- Triggering a focus nudge based on the window being closed forces us
  to make a guess / assumption about the viewport state (e.g. assuming that, if
  the window being closed was not the first, then we have a need to
  nudge the focus), while we could just recheck the current state
  instead (before deciding if we should take an action). *To be fair, the
  assumption was fairly safe but still slightly inelegant logic wise*.
- Only checking is the closed window was first in the viewport is
  insufficient. This doesn't take into account the case where the window
  being closed was between two existing window, unconditionally (and
  therefore unexpectedly sometimes) shifting the focus to the right.

This new logic now checks where the newly focused window is in the
viewport after a window got closed, and trigger a nudge if it's
positioned last. This is more elegant and robust because:

- We're not guessing / assuming the viewport state anymore, we actually
  check where the newly focused window is to determine if a focus nudge
  if needed or not.
- We can now reliably trigger a nudge focus only if needed (basically if
  the newly focused window is positioned last in the viewport), avoiding
  eventual arbitrary focus shift in some cases (basically leaving niri
  apply default focus shift in any other cases).

Also rename the related module to "reclaim_space" for consistency.
